project/requirements.txt
Django==1.9
Jinja2==2.7.2
Markdown==2.6.8
MarkupSafe==0.18
PAM==0.4.2
Pillow==2.3.0
Pygments==1.6
Sphinx==1.2.2
Twisted-Core==13.2.0
Twisted-Web==13.2.0
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
defer==1.0.6
dirspec==13.10
django-dajax==0.9.2
django-dajaxice==0.7
django-filter==1.0.1
djangorestframework==3.6.2
docutils==0.11
duplicity==0.6.23
gyp==0.1
html5lib==0.999
httplib2==0.8
lockfile==0.8
lxml==3.3.3
oauthlib==0.6.1
oneconf==0.3.7.14.04.1
pbr==2.0.0
pexpect==3.1
piston-mini-client==0.7.5
psycopg2==2.4.5
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pygraphviz==1.2
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
pyxdg==0.25
reportlab==3.0
requests==2.2.1
roman==2.0.0
sessioninstaller==0.0.0
six==1.10.0
software-center-aptd-plugins==0.0.0
stevedore==1.21.0
system-service==0.1.6
unity-lens-photos==1.0
urllib3==1.7.1
virtualenv==1.11.4
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wheel==0.24.0
wsgiref==0.1.2
xdiagnose==3.6.3build2
zope.interface==4.0.5
